+++ packages/browserslist-config-anolilab/packem.config.ts
@@ -12,5 +12,6 @@ export default defineConfig({
             path: "./LICENSE.md",
         },
     },
+    runtime: "node",
     transformer,
 });