+++ packages/eslint-config/packem.config.ts
@@ -1,6 +1,4 @@
-// eslint-disable-next-line import/no-extraneous-dependencies
 import { defineConfig } from "@visulima/packem/config";
-// eslint-disable-next-line import/no-extraneous-dependencies
 import transformer from "@visulima/packem/transformer/esbuild";
 
 export default defineConfig({
@@ -14,6 +12,7 @@ export default defineConfig({
             path: "./LICENSE.md",
         },
     },
+    runtime: "node",
     transformer,
     validation: {
         dependencies: {