+++ packages/textlint-config/packem.config.ts
@@ -11,6 +11,7 @@ export default defineConfig({
             path: "./LICENSE.md",
         },
     },
+    runtime: "node",
     transformer,
     validation: {
         packageJson: {